requirements.txt
# Python 3.13.2 호환 패키지 목록

# Claude AI API
anthropic>=0.39.0

# 이미지 처리
Pillow>=11.0.0

# HTTP 요청
requests>=2.32.0

# 환경 변수 관리
python-dotenv>=1.0.0

# 비디오 생성
moviepy>=1.0.3
decorator>=5.1.1

# OpenAI API (TTS)
openai>=1.57.0

# 오디오 분석
librosa>=0.10.2
soundfile>=0.12.1
numpy>=2.0.0
numba>=0.60.0

# 기타 의존성 (moviepy, librosa가 필요로 함)
scipy>=1.14.0
scikit-learn>=1.5.0
matplotlib>=3.9.0
imageio>=2.36.0
imageio-ffmpeg>=0.5.1
audioread>=3.0.1
soxr>=0.5.0
lazy-loader>=0.4

# 데이터베이스 (기본 포함되지만 명시)
# sqlite3는 Python 표준 라이브러리에 포함
